+++ natural_language_processing/fugumt-ja-en/fugumt-ja-en.cpp
@@ -421,7 +421,7 @@ static int recognize_from_text(AILIANetwork* encoder_net, AILIANetwork* decoder_
     std::vector<float> encoder_input_ids(tokens.size());
 	std::vector<float> attention_mask(tokens.size());
 	PRINT_OUT("Input Tokens :\n");
-	for (int i = 0; i < tokens.size(); i++){
+	for (size_t i = 0; i < tokens.size(); i++){
 		encoder_input_ids[i] = (float)tokens[i];
 		attention_mask[i] = 1;
 		PRINT_OUT("%d ", (int)encoder_input_ids[i]);
@@ -489,7 +489,7 @@ static int recognize_from_text(AILIANetwork* encoder_net, AILIANetwork* decoder_
 
         float prob = -INFINITY;
 		int arg_max = 0;
-		for (int i = 0; i < logits.size(); i++){
+		for (size_t i = 0; i < logits.size(); i++){
 			//PRINT_OUT("%f ", logits[i]);
 			if (prob < logits[i]){
 				prob = logits[i];
@@ -516,7 +516,7 @@ static int recognize_from_text(AILIANetwork* encoder_net, AILIANetwork* decoder_
 	PRINT_OUT("Output : %s\n",text.c_str());
 
 	PRINT_OUT("Output Tokens :\n");
-	for (int i = 0; i < tokens.size(); i++){
+	for (size_t i = 0; i < tokens.size(); i++){
 		PRINT_OUT("%d ", tokens[i]);
 	}
 	PRINT_OUT("\n");
