+++ util/wave_writer.cpp
@@ -95,7 +95,7 @@ void write_wave_file(const char *path, std::vector<float> data, int sampling_rat
 	}
 	int channel_n = 1;
 	write_header(fp, data.size(), channel_n, sampling_rate);
-	for (int i = 0; i < data.size(); i++){
+	for (size_t i = 0; i < data.size(); i++){
 		int pcm = data[i] * 32767;
 		short pcm_16bit = std::max(std::min(pcm, 32767), -32768);
 		fwrite(&pcm_16bit,2,1,fp);
